research_timeline_backend/requirements.txt
fastapi
uvicorn
requests
openai
semanticscholar
crossrefapi
python-dotenv
pydantic
scikit-learn
networkx
numpy
python-louvain 
community
scipy